Add scopes to OIDC configuration
This allows the CLI to retrieve custom scopes from the provider for use
when performing the oauth exchange to generate a token. Custom scopes
are useful, for example, when using the OIDC provider with Dex and
Github in that it allows the preferred_username or
federated_claims.user_id fields to be returned with the token for use in
SSH certificate templates (ie. to identify a Github user via
`extensions:login@github.com`).